# feat: use native crypto.randomUUID when available