--- a/src/native-browser.js
+++ b/src/native-browser.js
@@ -1,0 +1,4 @@
+const randomUUID =
+  typeof crypto !== 'undefined' && crypto.randomUUID && crypto.randomUUID.bind(crypto);
+
+export default { randomUUID };