--- a/src/native.js
+++ b/src/native.js
@@ -1,0 +1,3 @@
+import crypto from 'crypto';
+
+export default { randomUUID: crypto.randomUUID };