--- a/src/v4.js
+++ b/src/v4.js
@@ -1,8 +1,13 @@
+import native from './native.js';
 import rng from './rng.js';
 import stringify from './stringify.js';
 
 function v4(options, buf, offset) {
   options = options || {};
+
+  if (native.randomUUID && buf == null && options.random == null && options.rng == null) {
+    return native.randomUUID();
+  }
 
   const rnds = options.random || (options.rng || rng)();