*Before:*

```
uuidv4() x 2,232,325 ops/sec ±0.12% (95 runs sampled)
uuidv4() fill existing array x 8,398,262 ops/sec ±0.16% (97 runs sampled)
```

*After:*

```text
uuidv4() x 25,076,827 ops/sec ±1.93% (96 runs sampled)
uuidv4() fill existing array x 8,407,653 ops/sec ±0.21% (97 runs sampled)
```

This uses the native `crypto.randomUUID` if available, which results in more than an **11x** speedup on Node.js v16.11.1 on my M1 MacBook Air. All this while only increasing the bundle size with **34 bytes**!

<img width="710" alt="Screenshot 2021-11-30 at 09 46 51" src="https://user-images.githubusercontent.com/189580/144014610-f885cb84-7320-4fd8-b2ef-30155e3c5a22.png">


The patch implements my changes described here: https://github.com/uuidjs/uuid/pull/555#issuecomment-770720130

Closes #555