@LinusU 2021-11-30
@broofa I've adressed your comment and also updated the original post to reflect the new values (same perf, even smaller bundle size!)